requirements.txt
accelerate==1.2.1
TTS==0.22.0
tqdm==4.67.1
matplotlib==3.8.4
transformers==4.47.1
tokenizers==0.21.0
sherpa-onnx==1.10.38
sounddevice==0.5.1
soundfile==0.13.0
playsound==1.3.0

# Panel 相关
fastapi==0.115.6
uvicorn==0.34.0

# For Linux
torch==2.5.1; sys_platform == 'linux'
# For Windows
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1; sys_platform == 'win32'

torchaudio==2.5.1

# Controller相关
fastapi==0.115.6
uvicorn==0.34
pyjwt==2.10
tomlkit==0.13
sqlalchemy==2.0.37
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.0
python-multipart==0.0.20
bcrypt==4.2.0